+++ src/pages/app.tsx
@@ -3,8 +3,7 @@ import { useEffect } from "react";
 import BasicLayout from "../components/layouts/basic";
 import { AppFooter } from "../components/molecules/AppFooter";
 import { OrdersHeader } from "../components/molecules/headers/OrdersHeader";
-import { FAKE_ORDERS, getOrdersAsync, LINE_ITEM_CONTENT, LINE_ITEM_HEADER } from "../store/orderStore";
-import OrderStore from "../store/orderStore2";
+import OrderStore, { LINE_ITEM_CONTENT, LINE_ITEM_HEADER } from "../store/orderStore";
 import { useState, State } from '@hookstate/core';
 import { Router } from "@reach/router";
 import { store } from "../store/store";
