+++ src/startup.ts
@@ -1,5 +1,5 @@
 import UserStore from "./store/userStore"
-import OrderStore from "./store/orderStore2"
+import OrderStore from "./store/orderStore"
 
 export const startup = async () => {
   await UserStore.init()
